+++ test/modules.ts
@@ -28,6 +28,7 @@ import '@luma.gl/experimental/test';
 import '@luma.gl/arrow/test';
 import '@luma.gl/gpgpu/test';
 import '@luma.gl/text/test';
+import '@luma.gl/gpgpu/test';
 
 // DEPRECATED TESTS
 import '@luma.gl/constants/test';